+++ Origin/Camera.h
@@ -38,7 +38,7 @@ namespace illidan
 		Camera& operator=(const Camera& that);
 		virtual ~Camera();
 
-		void Update(float delta);
+		void Update(unsigned int delta);
 
 	//Virtual
 		void OnRButtonDown(WPARAM wParam, LPARAM lParam);
